requirements.txt
matrix-nio==0.24.0
markdown==3.6
openai==1.16.2
anthropic==0.23.1
pillow==10.3.0
sydney.py
git+https://git.evulid.cc/cyberes/bison.git
